Use haproxy 1.9.7 only
We're publishing our own haproxy 1.9.7 packages now, and we want to use
those instead of the ancient distribution package or our own 2.0-to-be
packages.

This introduces an overridable haproxy_package_version setting that
works much like postgres_package_version, but doesn't have a configure
option to set the value (i.e., it can be set only in config.yml).