Run bdr_camo_client verification on all instances in bdr_node_group
Until now, we involved only three instances in this test: ping and pong
(partners in the selected camo_node_group) and tester_instance, which
runs ``bdr_camo_client -h ping,pong``.

After bdr_camo_client finished, we then ran the verification query from
tester_instance against ping. Instead, we now run that query on all of
the instances in bdr_node_group, which also means we need to wait for
replication to settle on the instances first.

(Aside: I tried to set failures as a block var instead of a task var on
the contained include_tasks, and ended up looping over some sort of ssh
command output, where "failure" was set to stderr_lines instead of none.
Changing it to a task var worked around the problem.)