Always save bdr_camo_client output
Before this change, we wrote the registered stdout to a local file, but
that worked only if the client terminated properly. If the async_status
task failed, we would lose the output; but that's the situation in which
we need the output the most, so we save stdout/stderr to a file instead.

(We can't just ignore_errors here, because it's an async task, rather
than the way we run "repmgr check", for example.)