Create an haproxy role
We configure haproxy to use "pgsql-check" to check if the backend
postgres server(s) are alive. This just uses a few lines of code to
connect directly (i.e., without libpq) to Postgres and verify that it
receives an authentication request from the server.

Unfortunately, that leaves the server logs full of messages like this:

    FATAL:  password authentication failed for user "haproxy"
    DETAIL:  Role "haproxy" does not exist.

We don't actually want haproxy to be able to connect to Postgres as a
new user, but the only way to silence these errors is to create a new
role and assign it a password (without a password, the log spam just
changes, it doesn't go away).

Fortunately, we can create the role with "NOLOGIN", so that we don't
have a new user and attendant security concerns to deal with. Note that
haproxy offers no way to specify a password (it doesn't respond to the
authentication request, it just closes the connection immediately), and
the role wouldn't be allowed to login anyway.