Update camo configuration to use bdr_node_name
Now that we allow bdr_node_name to be set for an instance, we have to
use that name in BDR/CAMO configuration when required.

Rather than change the meaning of bdr_node_camo_partner (from the name
of an instance, as it is now), we just look up the referred instance's
bdr_node_name at the time of configuration.