Add basic haproxy tests
This just checks that we can connect to Postgres through haproxy, and
that the frontend/backend configuration looks sensible.

We also introduce a new ``|from_csv(column_names)`` filter that takes a
CSV line and a list of column names and returns a hash mapping the names
to the values from the line.