Make the network partition failure more resilient
If you apply the partition failure to an instance and rerun the test, we
cope sensibly with that situation now (in that we won't end up restoring
the partition rules after the partition ends). Also, if a restore fails
at some point, the subsequent application of the failure will behave
sensibly anyway.