Revise default postgres/src build configuration
On the basis that you're most likely trying to test something if you're
building from source, it is sensible to add --enable-debug and friends
by default (because you can't get debugging symbols any other way, and
having to go back and rebuild after something breaks is annoying).

The defaults are in postgres_configure_env and postgres_configure_opts;
they can be augmented with the settings in postgres_extra_configure_env
and postgres_extra_configure_opts. Existing clusters that specify the
value of postgres_configure_xxx will continue to work unmodified (but
without the new defaults, obviously).

If you're building from source for performance testing and need to build
without -g/-O0 etc., it's easy to override the defaults:

    cluster_vars:
        postgres_extra_configure_env:
            CFLAGS: "-O3"
        postgres_extra_configure_opts:
            - --disable-debug
            - --disable-cassert

(This is preferable to overriding postgres_configure_opts altogether.)