Remove unused 2ndQuadrant repository configuration files
With this change, we not only configure each repository mentioned in the
tpa_2q_repositories list, we also remove the configuration files for any
other repositories *not* mentioned in tpa_2q_repositories.

For example, if you were to deploy a cluster with

    tpa_2q_repositories:
        - products/2ndqpostgres/release

then edit config.yml to change the above to

    tpa_2q_repositories:
        - ci-spool/2ndqpostgres/snapshot

and re-deploy, the system would earlier end up with *both* repositories
configured, because the second deploy run knew nothing about the earlier
configuration. Now the 2ndquadrant-* files in /etc/apt/sources.list.d or
/etc/yum.repos.d not corresponding to entries in tpa_2q_repositories are
removed forthwith.

(This eliminates a source of confusion about selected package versions
that has happened multiple times in the past.)