Use python3 to acquire the barman lock if python2 doesn't work
With the release of Barman 2.8, the python-barman library has been
replaced with the python3-barman library. The lockfile script works
unmodified with python3, apart from the "#!/usr/bin/env python" line.
When run under python2, the import from barman.lockfile fails.

Instead of trying to detect which interpreter to use (note: both python2
and python3 will be installed and working at this point), we just try to
use python2 (as before) and then retry with python3 if the first attempt
fails. This may qualify as a dirty hack, but at least it's a small one.