Add become_user for barman lockfile tasks in force-archive
We should do barman-related tasks as the barman user, even when we can
get away with it otherwise.

We rename /tmp/lock to /tmp/lock.py so that we don't end up having some
problem trying to overwrite an existing root-owned file later (not that
its contents have changed, so that shouldn't normally happen). Arguably
that script should have been called lock.py to begin with anyway, so
it's a fortunate circumstance.