Fix logic error in handling tpa_2q_repositories
There are three possible cases:

1. If tpa_2q_repositories is not defined, we want to install
   dl/default/release.
2. If tpa_2q_repositories is explicitly set to [], we do not want to
   install any 2ndQuadrant repositories.
3. If tpa_2q_repositories is set to a list of repositories, we want to
   install dl/default/release and those repositories.

So the case we want to skip is when tpa_2q_repositories is defined AND
empty. The complement of that condition is when it's NOT defined OR not
empty. That's what the earlier condition expressed:

    tpa_2q_repositories|default('undefined') != []

In simplifying the condition (which I considered to be somewhat obscure,
probably correctly), I mistakenly left out the "is not defined" case.