Force pgdata_initialised=false for {start,stop}-postgres
We don't want to run cluster_discovery in either case (since it will
fail if Postgres is not running). Earlier, we included init.yml from the
command playbooks _without_ setting pgdata_initialised, but now we have
to explicitly set it to false to suppress the setting in roles/facts.

It's possible we need a separate toggle to suppress cluster_discovery
from postgres/facts, but there's no clear need to do so right now.