Add --hostnames-unsorted option to tpaexec configure
If you specify --hostnames-unsorted, hostnames will be used in the order
that they're found. This is useful especially for testing setups, where
you can statically assign understandable names for clusters to make it
easier to read and interpret test output.

Meanwhile, ``--hostnames-sorted-by`` must be used with the ``--a="b"``
syntax because there's no way to make argparse correctly interpret
"--dictionary-order" as an option value instead of an option.

References RM7483