Return only a single basedir from MinimalLoader.get_basedir
Using {{ lookup('pipe', …) }} in an --overrides-from file resulted in
errors like this:

    ansible.errors.AnsibleError: An unhandled exception occurred while
    running the lookup plugin 'pipe'. Error was a <type 'exceptions.TypeError'>,
    original message: coercing to Unicode: need string or buffer, list found

This is because plugins/lookup/pipe.py called self._loader.get_basedir()
and expected a single string in return. We want to maintain an array of
directories to implement our template fallback/override logic, but we
shouldn't return the whole list.

Thanks to Craig for the report.