Replace "{bdr_node_ids[n]}" in replication set names with the right id
BDR autoscaling uses replication sets with names like "shardset_ab_cd",
where ab and cd are the BDR node ids of the instances that subscribe to
the replication set (e.g., shardset_498589100_1465407329). Note that the
BDR node id is different from the (serial) node ids used in tpaexec. The
values are not known at configure time, so they have to be filled in at
deploy time.