Invoke postgres/facts from init only if postgres is running
Now that we detect and set pgdata_initialised in minimal_postgres_setup,
there are some new problematic corner cases that we have to cope with.
Since cluster_discovery will fail if it cannot connect to Postgres, we
should invoke it only if we have a reasonable chance to connect, which
means that the postgres user must exist (so that become_user works) and
Postgres must be running (so that the connection works).

It is not necessarily an error for these conditions to not be met early
during deployment, e.g., during rehydration.

We make minimal_postgres_setup set pgdata_user (UNKNOWN or a username)
and postgres_running (true or false) and test those conditions before
invoking postgres/facts from roles/init.

Aside: by now, we've taken a step firmly in the direction of detecting
an existing PGDATA and adapting our default settings. We did so only for
postgres_version already, but we can now do the same for postgres_user,
postgres_group, and so on. We could make minimal_postgres_setup detect
and set more such pgdata_xxx facts, and make facts use these values to
set defaults, or complain if incompatible values are explicitly set.