Adjust shared_preload_libraries ordering if required
If bdr is mentioned before either of pglogical or postgres_fdw, we need
to move it to the end of the list. We should not ordinarily encounter a
problem like this in the wild, because we always generate the correct
order to begin with.

However, if we end up with an ordering problem due to an implementation
bug (ahem), we can now fix it without much of a fuss.