Add *_client_dsn_attributes vars for SSL (and other) connection parameters
These vars are expected to be set under cluster_vars as follows:

    cluster_vars:
        postgres_client_dsn_attributes: >-
            sslmode=verify-full sslcert=… sslkey=… sslrootcert=…
        barman_client_dsn_attributes: >-
            sslmode=verify-full …
        repmgr_client_dsn_attributes: >-
            sslmode=verify-full …
        bdr_client_dsn_attributes: >-
            sslmode=verify-full …

Note that although it is _possible_ to set these for an instance, this
is very likely to break things. For example, the local_dsn passed to
bdr.create_node() MUST also work from other instances; likewise the
conninfo specified in repmgr.conf.

If you set, say, ``sslcert=x`` on one instance and ``sslcert=y`` on
another instance, and latter node will still use ``sslcert=x`` when
connecting to the former. If that certificate is not available, the
connection attempt will fail.

Use with caution.