Translate streaming_from via hostname_aliases, if possible
You can now declare additional hostnames/addresses for an instance:

    instances:
    - Name: main.host.name
      …
      vars:
        hostname_aliases:
          other.host.name: aa.bb.cc.dd
          yet.another.name: bb.cc.dd.ee

If we find a hostname in primary_conninfo that doesn't correspond to an
inventory_hostname we recognise, we can try to translate it based on the
hostname_aliases of the instance. We don't generate that situation, but
if recovery.conf is edited by hand to specify a different name for the
same host, we can adapt to that situation.