Make promote_command.sh rebuild pgbouncer.databases.ini
If repmgr_redirect_pgbouncer is true, we generate some extra lines in
promote_command.sh to update /etc/pgbouncer/pgbouncer.databases.ini on
every pgbouncer instance and reload pgbouncer.

This is experimental and incomplete. For example, no attempt is made to
ensure that the pgbouncer instance in question was pointing to the node
that was primary before the promotion, so in complex topologies it will
cause havoc. Also note that promote_command.sh is only executed for an
automatic failover, not a ``repmgr standby switchover`` invocation.

It might have been preferable to generate a pgbouncer.databases.xxx.ini
per postgres instance xxx and just have pgbouncer.databases.ini symlink
to the primary's version, but it's inconvenient to implement because of
the way pgbouncer_databases is defined in roles/facts. It may be worth
fixing this in the interests of making the switchover process easier.