Don't try to reuse elastic IPs on AWS
I enabled reuse_existing_ip_allowed by default when allocating elastic
IPs on AWS with some thought of minimising the total number of such IPs
we create, but this is not a sensible default in an environment where
multiple services draw EIPs from the same pool. Having a free EIP at all
may indicate a problem with the service that it was originally assigned
to, and it would be quite antisocial to opportunistically steal its IP.

It's better to clean up unused EIPs by some other means.