Don't use "nopurge" in the haproxy.cfg stick table definition
Per Shaun's analysis: «The current backend configuration of haproxy.cfg
intends for a node outage to revert to the alternate node in such a way
that reintroducing the original node does not result in connections
returning. This kind of sticky connection is unfortunately misconfigured
by using the nopurge flag in the stick-table definition. As a result,
returning a node to service results in traffic also being redirected to
that node.»

References RM7568