Associate a cluster's route tables only with *that cluster's* subnets
We inadvertently associated every new route table with *every subnet* in
the same region, whereas it should have been explicitly associated with
every subnet *in the cluster* in the region. We now do the latter.

(We always create route tables with the same routes, so earlier subnets
continued working even when associated with a new route table.)

Problem reported by Francesco.