Remove '%s'+!args test in postgresql_query
The test as written triggered on %%s, which was not actually an argument
string. Rather than make it ever more complicated in an attempt to keep
up, let's just remove it. We'll get list index lookup errors if a query
has %s and no args, but at least it'll be a genuine error.