Add linesinfile module
This module takes a file and a list of lines and ensures that the lines
exist somewhere in the file; if not, it appends each missing line to it.
The lines are literal lines (no regexps), and any insertions will be at
the end of the file. Despite the name, this module has only a fleeting
similarity to the much more complex lineinfile module.

This can be used to manage files like /etc/hosts without templating them
(thus replacing the whole contents, which can be troublesome especially
for platform bare) and without incurring the crippling overhead for
large clusters of running lineinfile in a loop.