Temporarily allow both Ansible 2.6 and 2.8
While I'm testing compatibility with 2ndQuadrant/ansible:dev/ansible-2.8
(and want others to test too), it's convenient to accept both versions.
The intention is to switch over to 2.8 in the near future, though, not
to maintain long-term compatibility with two versions.

Aside: this test should arguably check for the right version of
*2ndQuadrant* ansible, not just ansible.