Don't use the pythonN-psycopg2 package at all
On the one hand, it varies by distribution and version whether to use
python-psycopg2 or python2-psycopg2. On the other hand, some packages
available on PGDG are broken wrt. specific versions of Postgres.

For example, if you use 9.6 and install python2-psycopg2, it will fail
to link because it can't find PQencryptPasswordConn. If you install
psycopg2-binary with pip, it works.

Nobody would say it's a good idea to pip install packages into the
global Python installation, but what else can we do?