Allow absolute paths to template files
Ian noticed that we mishandled ``--overrides-from /some/file.yml``, in
that _tpaexec_get_template tried to open $basedir//some/file.yml, since
it only looked for templates under a specific set of basedirs.

This behaviour is fine for architecture- and platform-specific templates
that are used to generate config.yml, but ``--overrides-from`` is a bit
different, and it makes sense to accept absolute paths there.

Instead of introducing a separate mechanism to load "outside" templates,
we tweak _tpaexec_get_template a bit: if it can't find a template by the
normal rules and it has an absolute path to an existing template, it can
return its contents directly. So any invocation that worked before will
continue to work the same way, but we make an extra last-ditch effort
for absolute paths.

(We could also end up with an absolute path from an include within some
template, but that wouldn't have worked before this change anyway, so it
doesn't seem worth worrying about it here.)