Allow a custom haproxy_port to be set
Shaun and others have requested this in the past, but the feature should
be used with care. In the default BDR-Always-ON setup, where haproxy and
pgbouncer are both on the same instance, you should change haproxy_port
and pgbouncer_backend_port together. Or just leave them at the default
values.