Clear the stick table after setting the server to drain
If we don't do this, the stick table remains unmodified, and as far as
pgbouncer is concerned, the existing server works just fine. Only when
we clear the stick table does haproxy bother to reconnect to the other
server in the pool.

Setting the server to maint does close the existing connections, but we
can't do that straightaway because we need to be able to tell pgbouncer
to migrate the connections sensibly.