Allow updates to specific versions in pkg/update.yml
Earlier, we would always try to upgrade to the latest available package.
That worked as far as it went, but it did not allow you to specify new
versions to upgrade to. Now we use "latest" if no version is specified,
and "present" if any versions are specified.

This means that:

1. If you don't define any xxx_package_version (e.g., if you configure
   your cluster without any --xxx-package-version options), the update
   will try to install the latest available versions.

2. If you DID use --xxx-package-version (or define xxx_package_version
   in some other way), the update will probably not do anything, since
   the required packages are presumably already present.

3. If you _change_ the values of xxx_package_version (either by editing
   config.yml and re-provisioning, or with ``-e xxx_package_version=…``
   on the update command-line), then the package manager will make some
   desultory effort to satisfy your demands.

This means it's possible to break your cluster by specifying conflicting
version requirements (or under-specifying which versions you want). As a
matter of caution, you should either specify no versions or all of them.