Don't set net.ipv4.ip_forward by default
We don't need it unless we're configuring openvpn, and then only on the
server, so we should avoid always forcing it to 1.

Note that re-deploying with this change won't disable ip_forward on any
instances where it's already enabled (to avoid breaking anything). Also,
on systems where it's been turned off with a post-deploy hook (as shown
in the issue linked below), it'll remain off.

References: #11424