Ensure that a replica does not have max_worker_processes < the primary
Otherwise you get errors like this when you start the replica:

    FATAL:  hot standby is not possible because max_worker_processes = 6
    is a lower setting than on the master server (its value was 16)

The circumstances in which this could happen with the computed default
are convoluted (adding and removing instances from a cluster with many
disjoint replication pairs), but in any case, ensuring that the value
matches or exceeds the primary's setting is sensible.

If you set max_worker_processes explicitly (which was a way to work
around the problem, if you hit it), this change won't affect anything.

Delivered: #6035