Turn off via_haproxy by default
The test as written fails if it finds an haproxy instance whose backends
are not a CAMO pair, i.e., it always fails on non-CAMO clusters. That is
not acceptable for the default test—it should do something sensible on
whatever sort of cluster you run it on. It's fine to detect and skip
tests that it can't run, but it's not OK to fail.

The correct approach here would be to look for an haproxy instance whose
backends are a CAMO pair. If there is one, we can run the tests through
it. If not, we can run the tests against the CAMO pair directly. But I
don't have the time for that right now, so a bigger hammer it must be.