Remove ec2_availability_zone setting
I didn't realise that ec2.py published the instance's availability zone
as ec2_placement, so the two workarounds (parsing the instance meta-data
in metadata.yml, or retrieving it from describe_instance_status) weren't
necessary after all.

As it happens, we don't currently use ec2_availability_zone anywhere, so
we can just remove the code that set it.

It's possible that this change will break some code not in tpaexec or
tpaexec-contrib that depends on ec2_availabilty_zone being set, but it
would be better to fix that code to use ec2_placement than to default to
setting ec2_availability_zone here (to avoid any conflict with the ec2
inventory script in future).