Run wait_for_ssh/ec2_instance_status only from deploy.yml
Earlier, roles/platforms would run wait_for_ssh and wait to be able to
connect via ssh, and roles/platforms/aws would run ec2_instance_status
and wait for the instance reachability checks to pass. These are both
important tests to run just after ``tpaexec provision`` is completed,
but they're less important for other commands.

So now we introduce a ``wait_for_instances`` variable that controls this
behaviour, and set it to false by default and true only in deploy.yml.
So ``tpaexec deploy`` will behave the same as always, but ``tpaexec
test`` and other commands will skip the wait.

Custom commands that do want to test and wait can set wait_for_instances
to true when importing init.yml (as the deploy.yml playbooks now do).

(Aside: in an ever so slightly dubious related decision, we also skip
testing for user-data script failures when we're not waiting for the
instance. If there are commands that we _expect_ to run before deploy,
we might need to revisit this.)

(Aside²: Custom commands that don't want to wait for other things, like
roles/selftest, are still free to do their own thing without importing
init.yml at all.)