Add haproxy agent checks
The pgsql-check in haproxy tests only if it can connect to Postgres. If
you want to do more than that, for example to run a query and base your
decision on its result, you can configure an external check and haproxy
will execute the specified program for you.

External checks are given the address and port of a server to check, but
they have no access to haproxy internal state, such as the active server
or the status of earlier health checks for other servers. This makes it
difficult to implement complex health checks that assess the state of
the cluster as a whole.

We would really like to be able to run one or more checks independently
on several servers and then take some decision based on the cumulative
results. Doing this with external checks would mean having to maintain
state between concurrent executions (for different servers) and using
multiple haproxy stats socket commands to try to observe haproxy state.

You can run multiple checks for a server by defining multiple backends,
each with its own check, and defining the same set of servers under each
backend. You can then have a final active backend (i.e., with a frontend
to direct traffic to it) with a check that synthesises the results of
the checks defined under all the other inactive backends.

You can, however, configure a health check and an "agent" check to run
simultaneously on the same server under the same backend. So you can use
pgsql-check but also query the agent over a TCP connection and receive a
response that overrides the health check result. This saves one backend,
because the active backend can configure the basic pgsql-check as well
as an agent-check to synthesize check results from the other backends.

Finally, thanks to the Lua embedding API, we can register the agent as a
service that runs in haproxy itself. This gives the agent convenient and
fast access to haproxy internal state without having to repeatedly query
the haproxy stats socket and parse the results.

The configuration for such a system would look like this:

    global
        lua-load agent.lua
        external-check

    # Route Postgres connections to the active backend
    frontend fe
        bind 127.0.0.1:5432
        default_backend be

    # Route agent-checks to service defined in agent.lua, loaded above
    frontend agent
        bind 127.0.0.1:12345
        tcp-request content use-service lua.agent

    # Active backend to route requests from fe to Postgres servers s1…sN
    backend be
        stick-table type ip size 1
        stick on dst

        option pgsql-check user haproxy

        server s1 s1:5432 … check inter 1500 rise 5 fall 3 … agent-check agent-inter 1500 agent-addr 127.0.0.1 agent-port 12345 agent-send "s1\n" …
        server s2 s2:5432 … check inter 1500 rise 5 fall 3 … agent-check agent-inter 1500 agent-addr 127.0.0.1 agent-port 12345 agent-send "s2\n" …
        server s3 s3:5432 … check inter 1500 rise 5 fall 3 … agent-check agent-inter 1500 agent-addr 127.0.0.1 agent-port 12345 agent-send "s3\n" …

    # Backend defined only to execute check1
    backend be_check_xxx
        option external-check
        external-check command "/etc/haproxy/check-xxx"

        server s1 s1:5432 … check inter 1500 rise 5 fall 3 …
        server s2 s2:5432 … check inter 1500 rise 5 fall 3 …
        server s3 s3:5432 … check inter 1500 rise 5 fall 3 …

    # Backend defined only to execute check2
    backend be_check_yyy
        option external-check
        external-check command "/etc/haproxy/check-yyy"

        server s1 s1:5432 … check inter 2500 rise 2 fall 4 …
        server s2 s2:5432 … check inter 2500 rise 2 fall 4 …
        server s3 s3:5432 … check inter 2500 rise 2 fall 4 …
    …

(Yes, there's really no way to split up long lines in haproxy.cfg)

You can tune the inter/rise/fall configuration based on how expensive or
important each check is. There are internal agent-rise/fall values, but
they are pinned to 1 and are not configurable, so a response from the
agent will immediately override the server status.

With this configuration, we could register a service named "agent" that
looks at the external/health check results for servers s1…sN under each
backend and authoritatively decides the final status of the server. (We
would need to either hardcode the names/nomenclature of backends in the
agent, or also agent-send the names of each backend to check.)

One additional factor in this configuration is the stick-table/stick on
directives under the active backend. These "stick" a particular request
source address (127.0.0.1, because we always connect to haproxy through
pgbouncer running on the same host) to a particular server_id. HAproxy
will by default always send connections to the active server (s1 above)
and fall back to the first available server marked "backup" when s1 is
down, but we need the stick table so that we don't fall _back_ to s1 as
soon as it is back up again.

The immediate reason to do all of this is so that we can control servers
during the failover process. Specifically, we want CAMO servers to start
accepting connections after failover only after dealing with the pending
apply queue, so that clients see the up-to-date state after a failover.
It's better to have no active server for a brief while than to failover
to a server that's still catching up.

With review, testing, and bugfixes from Ian.

References: 11664