Add --os-version configure option
Now you can do

    tpaexec configure --platform vagrant --os RedHat --os-version 8

This works for Debian (8/jessie, 9/stretch, 10/buster), RedHat (7, 8),
and Ubuntu (16.04/xenial, 18.04/bionic, 20.04/focal) on Vagrant and AWS.
Each platform has to interpret the (label,version) provided to determine
the image name to use.

(Note: this does NOT allow you to use CentOS 8 on Docker. Yet.)

This is currently not formally documented, because it allows you to
configure a cluster with a distribution that isn't fully supported.