Make ``git clone`` tasks work on Docker
The problem was that access to private git repositories depended on
using ssh agent forwarding, but we didn't use ssh to access docker
containers, so it didn't work.

We solve this by setting ``ansible_connection: ssh`` for git clone tasks
(only), so that forwarding works as before. We also need to ensure that
ansible_ssh_host is set to the IP address, because otherwise there's no
way to translate the instance name to an address (see inventory.j2).