Avoid stupid warnings
During tpaexec provision, the docker_container task would issue these
warnings:

 [WARNING]: The value 22 (type int) in a string field was converted to u'22' (type string). If this does not look like what you expect, quote the entire value to ensure it does not change.
 [WARNING]: The value 5432 (type int) in a string field was converted to u'5432' (type string). If this does not look like what you expect, quote the entire value to ensure it does not change.
 [WARNING]: The value 6432 (type int) in a string field was converted to u'6432' (type string). If this does not look like what you expect, quote the entire value to ensure it does not change.

This is because the exposed_ports module parameter is declared as a
``type='list', elements='str'``. The warning is issued for any integer
port you specify, i.e., even for the example usage in the documentation
for the module.

Moral of the story: if you don't want your integers to be turned into
strings, make sure they are strings to begin with.