Allow postgres/pgpass to write selected entries to a given file
Earlier, this role was hardcoded to write entries for postgres/repmgr to
~postgres/.pgpass. Now the caller can optionally specify the location of
the .pgpass file, its owner and group (mode is fixed at 0600; libpq will
insist on that anyway), and the entries it should have.

    - include_role:
        name: postgres/pgpass
      vars:
        pgpassfile: /etc/haproxy/.pgpass
        pgpass_owner: root
        pgpass_group: root
        pgpass_users:
        - check_queue

This example will write a line for the check_queue user (only) to
/etc/haproxy/.pgpass. If any of these vars/settings are omitted, the
defaults remain the same as before.