Get rid of sudo-check-queue
Now that bdr.get_configured_camo_partner_of() can be executed by the
bdr_application role, we can get rid of the sudo to postgres and the
need to use superuser privileges to run the check, and use a new
unprivileged user that is granted the bdr_application role.

This needs a bit of juggling to ensure that we correctly create the
unprivileged check_queue user. Also, we use socket.gethostbyaddr()
instead of bdr.local_node_summary to compute "this_node", because
bdr_application can't access the view.

haproxy_backends_are_camo is renamed to
haproxy_backends_need_queue_checks