Make minimal_setup also return a "python" fact
This is either "python2" or "python3", depending on which interpreter we
should try to use on the target (which may not be installed yet).

We use python2 on debian/jessie because it has only Python 3.4 available
(not supported by Ansible, which required at least 3.5).

We use python2 on RHEL/CentOS 7 because they don't have working dnf, and
the yum module doesn't work with Python 3 because there's no RPM binding
package available. So we have to use python2 to use yum, even though we
do support the available Python 3.6.8.

Everywhere else, we can try to use python3.

(Note: nothing acts upon this fact yet.)