Install the preferred version of Python
We always used to install 2.7. Now we install python/python2.7/python3
depending on the platform and what python minimal_setup decided should
be used.

As an optimisation, we try to avoid running apt/yum if the interpreter
is already available. Likewise, we check if backports.ssl_match_hostname
is importable before trying to install it (for older versions of Python2
only). This matters less for apt than yum, which may spend a much longer
time downloading stuff before deciding there's nothing to do.