Remove force_basic_auth from repository installation
"We need to support https://user:pass@host/… key and repository URLs",
said the comment, but we have never actually needed that. Rather than
make it work for dnf (which we now use on RHEL/CentOS 8), we can just
remove it altogether.

We should revisit this if we ever really "need to support …".