Add preferred_python_version and change ansible_python_interpreter
You can now set ``preferred_python_version: python2`` or ``python3`` in
cluster_vars or instances[*].vars.

This setting declares a preference, which minimal_setup may choose to
disregard—we will only ever support python2 on jessie/centos7, and we
only support python3 on centos8. But stretch/buster/xenial/bionic can
all support both python2 and python3.

If unspecified, the default is python2, for backwards compatibility with
existing clusters. The default for newly-configured clusters is python3.

We set ansible_python_interpreter to ``/usr/bin/{{ python }}`` now. We
would prefer to set it to ``auto_silent``, but there's no way to tell
Ansible to look for a python2 or python3 interpreter: it'll just find
*some* python interpreter, regardless of your preference.

It so happens that installing the desired version of python will cause
/usr/bin/pythonN to work on the platforms we support, so we can avoid
interpreter detection altogether. (On the other hand, it takes some
effort to write literal ``{{ expressions }}`` to a file without any
template expansion!)

We go to some lengths to set a default for ansible_python_interpreter.
Since the ``python`` fact will be set only by minimal_setup, it will not
be available for adhoc tasks. We can declare a default of python2.7 for
backwards compatibility, but then centos8 won't work. So we end up with
``python|default(preferred_python_version|default('python2.7'))``, which
works for (a) playbooks that use minimal_setup to set python, (b) hosts
that have preferred_python_version set as a var, and (c) everyone else
that already works.