Don't try to install awscli
We've never been sure when it is necessary to install it, and now with
python3, the invocation fails because there's no virtualenv command by
default. There must surely be a way to fix that, but since this is the
only use of virtualenv and since we install awscli in user-data on EC2
instances anyway, it's easier to remove it altogether.