Avoid conflict with awscli over PyYAML version
We require PyYAML>=5.1, which is now resolved to 5.3; but awscli
requires PyYAML<5.3, so people have been reporting this error from
``tpaexec setup`` for some time now:

    ERROR: awscli 1.18.1 has requirement PyYAML<5.3,>=3.10, but you'll
    have pyyaml 5.3 which is incompatible.

Things appear to work despite this error, but we don't care about 5.3 in
particular, so we can adopt the same restriction as awscli and avoid the
worrying error message.

Thanks to Marco for the analysis and suggested fix.