Don't use an already-active virtualenv under pyenv
Under pyenv, it's quite normal to have some virtualenv active already,
but if we use it, we will later be confused about where ansible lives
when command -v points to .pyenv/shims/ansible and not tpa-venv/bin.

Rather than trying to cope with pyenv's special paths, we can just
always create our own venv if pyenv is active, and continue on as
before.

Thanks again to Marco for the discussion and suggested fix.