Remove postgresqlNN-devel from the default package list
Under RHEL/CentOS, installing postgresqlNN-devel for 2ndQPostgres from
the BDR 3.6.15 release incurs a dependency on llvm-toolset-7-clang, not
available from the default repositories. Under CentOS, installing the
centos-release-scl package enables the correct repository, but the
equivalent ``yum-config-manager --enable rhel-server-rhscl-7-rpms``
invocation did not work for me on RHEL.

There is work underway to remove the extra dependency. Until this is
completed, we can just not install the -devel package by default. This
is not a "nice" fix, but it avoids the problem for new clusters.

However, upgrades will not work unless the -devel package is removed
beforehand. Also, BDR-specific tests will not work unless the -devel
package is installed (because the bdr-test package requires it).