Enable pgdgNN and pgdgNN-updates-debuginfo repositories for PGDG
The postgresqlNN-debuginfo packages used to be available through the
pgdgNN repository, but they were moved to another repository without
warning. So now we have to enable multiple pgdgNN repositories and
disable all others.

(We try to do this more efficiently by fetching a list of entries we may
need to change the status of in one task, and looping over those entries
instead of either fetching or trying to predict the names of all entries
the file may contain.)

Unfortunately, this still doesn't work, because the server is missing
some directories like "rhel-7Server-x86_64" for RHEL.

https://download.postgresql.org/pub/repos/yum/debug/11/redhat/rhel-7Server-x86_64/repodata/repomd.xml: [Errno 14] HTTPS Error 404 - Not Found

References: 13944