Add docker_network setting to create user-defined docker networks
You can now set the following in config.yml:

    docker_networks:
    - name: xxx
    - name: yyy

These networks will be created during provision if they do not exist,
and removed unconditionally during deprovision.

This new setting is related to, but independent from, the existing
instances[*].networks setting. The former defines which networks we
should create. The latter defines which networks a container should be
attached to—and it stands to reason that the networks you create may
also feature here. Note that the syntax is different, though: one is
based on docker_network parameters, the other docker_container ones.

If a container is not in the default network (as it will not be, if you
specify user-defined networks in instances[*].networks), it will have an
empty NetworkSettings.IPAddress. In this case, we look for the IPAddress
from the first network and use that for private_ip instead.

Based on a proposal from Amruta in RM12941.